+++ packages/utils/theme-node/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.24](https://github.com/adevinta/spark/compare/@spark-ui/theme-utils-node@2.0.23...@spark-ui/theme-utils-node@2.0.24) (2023-03-15)
+
+**Note:** Version bump only for package @spark-ui/theme-utils-node
+
+## [2.0.23](https://github.com/adevinta/spark/compare/@spark-ui/theme-utils-node@2.0.22...@spark-ui/theme-utils-node@2.0.23) (2023-03-15)
+
+**Note:** Version bump only for package @spark-ui/theme-utils-node
+
+## [2.0.22](https://github.com/adevinta/spark/compare/@spark-ui/theme-utils-node@2.0.21...@spark-ui/theme-utils-node@2.0.22) (2023-03-14)
+
+**Note:** Version bump only for package @spark-ui/theme-utils-node
+
 ## [2.0.21](https://github.com/adevinta/spark/compare/@spark-ui/theme-utils-node@2.0.20...@spark-ui/theme-utils-node@2.0.21) (2023-03-13)
 
 **Note:** Version bump only for package @spark-ui/theme-utils-node